requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==3.0.4
channels_redis==4.2.1
charset-normalizer==3.3.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
daphne==3.0.2
dj-database-url==2.2.0
Django==5.1.1
django-allauth==64.2.1
django-extensions==3.2.3
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
fonttools==4.53.1
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
msgpack==1.1.0
numpy==2.1.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
wordcloud==1.9.3
zope.interface==7.2
